complete 64bit interrupts part
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/wsim@747 c0ef0dea-404d-0410-ad65-84cceb5f200a

+++ arch/msp430/msp430_alu.h
@@ -15,7 +15,7 @@ typedef int16_t mcu_register_t;
 
 struct msp430_alu_t {
   mcu_register_t   regs[MCU_REGISTERS];
-  uint32_t         interrupt_vector;
+  uint64_t         interrupt_vector;
   uint32_t         signal;
 
   uint64_t         insn_counter;
